investment-team: round-5 review — clean-code refactor + test split
Applies the full S1-S3 + M1-M6 batch from the round-5 review.

## Aggregator refactor

- **S1** ``_runtime_events_to_blockers`` was sort-then-filter. Restructured
  to filter → sort → map: events without a valid ``rule_id`` are dropped
  upfront, the survivors are numerically sorted, then each maps to a
  ``LikelyBlocker`` via the new ``_event_to_blocker`` helper. The
  ``_rule_sort_key`` defensive ``None`` branch becomes dead code and is
  removed — its sole caller now always passes a non-empty string.
- **S3** ``run_coverage_stage`` docstring undersold the control flow
  ("three-stage" vs. five actual phases with two early exits). Expanded
  to the accurate five-phase enumeration.

## Clean-code helpers

- **M1** ``_pick_category`` replaces ``dict.__getitem__`` magic with a
  named ``_category_rank`` function so the call reads as
  ``min(..., key=_category_rank)``.
- **M2** ``_fetched_universe`` + ``_longest_symbol_bars`` collapse into
  a single ``_summarize_market_data(market_data) -> (universe,
  longest_bars)``. One pass instead of two; named tuple return.
- **M3** Outcome interpretation extracted from ``_runtime_reexecute``
  into ``_interpret_probe_exec(probe_exec, rule_labels)``. The
  orchestration function now reads as "set up + run subprocess + handle
  exception"; the helper handles every observable post-exec state.
- **M4** ``_to_dataframes`` switched from ``pd.to_datetime(..., errors=
  "coerce")`` + ``.isna().any()`` (two passes, build-then-discard) to a
  ``try/except`` with ``errors="raise"`` (single pass, short-circuit on
  first bad date).
- **M5** OHLCVBar → DataFrame conversion lifted into a new
  ``_ohlcv_list_to_dataframe`` helper that uses ``pd.DataFrame.from_records``
  with explicit columns. Skips the per-bar ``pydantic.model_dump()`` cost
  via direct attribute access — meaningfully faster on multi-thousand-bar
  history.
- **M6** Runtime-stage token vocabulary centralised into module-level
  ``_RUNTIME_NOT_RUN`` / ``_RUNTIME_SKIPPED`` / ``_RUNTIME_FAILED`` /
  ``_RUNTIME_NO_FRAME`` constants. ``_summary_line``, ``_skip``, ``_fail``,
  ``_no_frame`` all reference the named constants — no more
  duplicated literal strings spread across docstrings.

## Test split (S2)

Original ``test_coverage_probe_orchestrator_stage.py`` was 1,242 LoC
mixing four distinct concerns. Split into:

- ``_coverage_probe_test_helpers.py`` — shared fixtures (``make_config``,
  ``make_spec``, ``make_flat_df``, ``make_diag``, ``make_exec_result``,
  ``make_report``, ``make_ohlcv_series``, ``unknown_indicator``,
  ``runtime_capable_code``, ``never_called_run_strategy_code``). Leading
  underscore keeps pytest from collecting it.
- ``test_coverage_probe_aggregator.py`` (208 LoC) — gate + merge + dedup
  + module invariant unit tests.
- ``test_coverage_probe_stage.py`` (622 LoC) — ``run_coverage_stage``
  end-to-end: static short-circuit, runtime gating, B1/B2/Q4 runtime
  outcomes, OHLCVBar conversion, determinism, LLM-free guarantee.
- ``test_coverage_probe_helpers.py`` (96 LoC) — ``_to_dataframes`` direct
  tests (R2/R3) plus a new ``test_to_dataframes_passes_dataframe_inputs
  _through_by_identity`` to cover the helper's identity contract.
- ``test_coverage_probe_orchestrator_stage.py`` (151 LoC, slimmed) —
  ``_maybe_attach_coverage_report`` tests + AST source-level
  regression. The file's existing name is preserved so review history
  on PR #475 keeps a stable anchor.

Each test file is now well under the "consider splitting" threshold and
each addresses one concern. Test count: 43 (was 42; new helper-identity
test).

## Results

- ``pytest agents/investment_team/tests/test_coverage_probe_*.py`` — 43/43 pass
- Regression on probe + orchestrator suites: 321/321 pass
- ``ruff check`` + ``ruff format --check`` clean on all 7 touched files
